Added support for different sheet types
The program checks the ClassAndLevel element and changes the maximum
feats and class features allowed accordingly. The pdf formats the
element strangely however (it adds in formatting information) which
makes it hard to read. I'll see if I can fix the pdf so it doesn't
format the element.

Also changed a couple of comments that didn't make sense.