Implemented default filename code
Standard format is "[charactername]_lv[level].xml" with a zero in front
of the level if the number is below 10 e.g. "Braxus_lv05.xml". One thing
I'd like to change is not display the whole filepath, instead showing
just the file name itself. This will require parsing the sourceFile
variable. I'll do that at some point but it's not a priority.