Finished implementing attack and damage support
Will look at colouring the HP bar depending on current HP total next.
After that, the delay system. After that, will add functionality to the
"move up" and "move down" buttons (this will require some R&D on the
ordering of items in a collection). Finally I'll implement effects (+1
to all attack rolls for example) along with remaining rounds of effect
that are decremented each time the round number is incremented.