Implemented "Wound Combatant" feature
Ctrl + W (wound) now opens a dialog for dealing damage to a specific
combatant. This allows damage to be dealt without changing the selected
combatant index in the initiative list. Also added Ctrl + N shortcut
(New round) which increments the round number and automatically selects
the first combatant in the initative list. Began R&D on overloading the
"DrawItem" event for ListBox in order to change the colour of
combatant's names in the initiative list depending on their current HP
status (black for healthy, yellow for under half and red for under a
quarter).