Implemented colour coding in initiative list
Added a feature that means the colour of the combatant's name in the
list indicates their hp. Orange means <= half, red means <= a quarter
(need to do this for the wound, delay and dying lists too). Also added
effects group to bottom right although it is unpopulated at this point.