Began implementing "Save Battle" functionality
Program saves data in .btl files. So far all it does is store the
initiative list. The next step is the delay list and finally the dead
and dying list. After that I'll work on loading the data back in (that
should be interesting).

Oh! Quick note: must tweak effect duration reduction code. I believe it
only reduces the effect duration of effects on combatants in the normal
initiative list. It needs to do so with the other two lists as well.