Finished Save/Load implementation
It's pretty goOOOod. I'd like to create prompts for whenever the user
tries to exit or open a new file without saving existing changes. That
will require planting checks wherever data is changed (indices,
combatant stats, and round number) and then flagging them up. The flag
will be reset whenever the file is saved. I'll either create and hook up
a "dataChange" event to fire (with an event handler that flags the
changes) or pop in a method call to a method that does the flagging.
Even though it's more challenging I'm leaning toward the event solution
as it has panache and I think I'll learn a lot from it.