Finished Player tab prototype
Tweaked dialogs and added a couple of new features, one being a save
shortcut (no dialog needed if you just click "Save" and have a file
loaded), the other being the option to append combatants to a battle.
That way, an organised DM can prepare an encounter file and load in the
enemies when necessary instead of on the fly like a common troll.